+++ examples/16_network/pet_segmentation_anim.py
@@ -27,7 +27,7 @@ class VideoAnimation(FuncAnimation):
 
     def save(self, *args, **kwargs):
         if args[0].endswith("gif"):
-            # In this case gif is use to create thumbnail which are not use but consume same time than video
+            # In this case gif is used to create thumbnail which is not used but consume same time than video
             # So we create an empty file, to save time
             with open(args[0], "w") as _:
                 pass
@@ -96,8 +96,7 @@ def update(i_frame):
 
     indices_frames = INDICES[i_frame]
     mappable_CONTOUR.set_data(
-        e.contour_lon_e[indices_frames],
-        e.contour_lat_e[indices_frames],
+        e.contour_lon_e[indices_frames], e.contour_lat_e[indices_frames],
     )
     mappable_CONTOUR.set_color(cmap.colors[tr[indices_frames] % len(cmap.colors)])
     return (mappable_tracks,)